# docs/swagger.yaml
definitions:
...
  controller.phoneInput:
    properties:
      brand_id:
        type: integer
      image_url:
        type: string
      model:
        type: string
      price:
        type: integer
      release_date:
        type: string
    type: object
  controller.profileInput:
    properties:
      biodata:
        type: string
      birthday:
        type: string
      full_name:
        type: string
      image_url:
        type: string
    type: object
...
  controller.specificationInput:
    properties:
      additional_feature:
        type: string
      battery:
        type: string
      camera:
        type: integer
      memory:
        type: integer
      network:
        type: string
      operating_system:
        type: string
      storage:
        type: integer
    required:
    - battery
    - camera
    - memory
    - network
    - operating_system
    - storage
    type: object
...
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      review_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Phone:
    properties:
      brand_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      model:
        type: string
      price:
        type: integer
      release_date:
        type: string
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      specification:
        items:
          $ref: '#/definitions/models.Specification'
        type: array
      updated_at:
        type: string
    type: object
  models.Profile:
    properties:
      biodata:
        type: string
      birthday:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: integer
      image_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Review:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      phone_id:
        type: integer
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Specification:
    properties:
      additional_feature:
        type: string
      battery:
        type: string
      camera:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      memory:
        type: integer
      network:
        type: string
      operating_system:
        type: string
      phone_id:
        type: integer
      storage:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /phones:
    get:
      description: Get a list of Phone.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Phone'
            type: array
      summary: Get all Phones data. (PUBLIC)
      tags:
      - Phones
    post:
      description: Creating a new Phone data, only account with role admin can accsess
        this route
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: example JSON body to create a new Phone, sample release_data
          format = 2023-09-22T00:00:00Z
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.phoneInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
      security:
      - BearerToken: []
      summary: Create New Phone (ADMIN ONLY)
      tags:
      - Phones
  /phones/{id}:
    delete:
      description: Delete a Phone by id, only account with role admin can access this
        route
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete Phone by id . (ADMIN ONLY)
      tags:
      - Phones
    get:
      description: Get a Phone by id.
      parameters:
      - description: phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
      summary: Get Phone. (PUBLIC)
      tags:
      - Phones
    put:
      description: Update Phone data by id, only account with role admin can access
        this route
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      - description: Example JSON body to update Phone data, sample release_data format
          = 2023-09-22T00:00:00Z
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.phoneInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Phone'
      security:
      - BearerToken: []
      summary: Update Phone data. (ADMIN ONLY)
      tags:
      - Phones
...
  /phones/{id}/specification:
    get:
      description: Get all Phones data by phone id.
      parameters:
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Phone'
            type: array
      summary: Get specification data by Phone id. (PUBLIC)
      tags:
      - Phones
    post:
      description: Creating a specification data for phone, only admin can access
        this route
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Phone id
        in: path
        name: id
        required: true
        type: string
      - description: example JSON body to create a specification for phone
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.specificationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Specification'
      security:
      - BearerToken: []
      summary: Create Specification for phone (ADMIN ONLY)
      tags:
      - Phones
    put:
      description: Creating a specification data for phone, only admin can access
        this route
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: phone id
        in: path
        name: id
        required: true
        type: string
      - description: example JSON body to update a specification for phone
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.specificationInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Specification'
      security:
      - BearerToken: []
      summary: Update Specification for phone (ADMIN ONLY)
      tags:
      - Phones
  /profiles:
    post:
      description: Creating a profile data for user, user ID is taken from JWT Token
        so only acount's owner can create the profile
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: example JSON body to create a profile for user, user_id is taken
          from the authorization token
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.profileInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Profile'
      security:
      - BearerToken: []
      summary: Create Profile for user
      tags:
      - Profiles
    put:
      description: Updating a profile data for user, user ID is taken from JWT Token
        so only acount's owner can update the profile
      parameters:
      - description: 'Authorization : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: example JSON body to update a profile for user, user_id is taken
          from the authorization token
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.profileInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Profile'
      security:
      - BearerToken: []
      summary: Update Profile for user
      tags:
      - Profiles
...
swagger: "2.0"